Improve handling of boolean attributes
Those attributes where their presence signifies "true" must not be
rendered to convey "false". For example: "required" will render as ,
`required="required". But this happens regardless of whether the boolean
is true or false. These changes rectify this situation by not rendering
the attribute in case the value is "false".